// src/Interfaces/CustomerData.ts
export interface CustomerData {
  fullName: string
  email: string
  phone: string
  document: string
  zipCode: string
  street: string
  number: string
  complement?: string
  neighborhood: string
  city: string
  state: string
  creditCard: string
  creditCardHolder: string
  creditCardExpiration: string
  creditCardSecurityCode: string
}